+++ R/vignette.R
@@ -1,12 +1,15 @@
-#' Create a vignette
+#' Create a vignette or article.
 #'
-#' Performs general set up for vignettes and initializes a new individual
-#' vignette:
-#' * Adds needed packages to `DESCRIPTION`
-#' * Adds `inst/doc` to `.gitignore` so built vignettes aren't tracked
-#' * Creates a new draft `.Rmd` vignette in `vignettes/` and, if possible,
-#' opens it for editing
+#' Creates new vignette/article in `vignettes/`. Article are a special
+#' type of vignette that appear on pkgdown websites, but are not included
+#' in the package itself (because they are added to `.Rbuildignore`
+#' automatically).
 #'
+#' @section General setup:
+#' * Adds needed packages to `DESCRIPTION`.
+#' * Adds `inst/doc` to `.gitignore` so built vignettes aren't tracked.
+#' * Adds `vignettes/*.html` and `vignettes/*.R` to `.gitignore` so
+#'   you never accidental track rendered vignettes.
 #' @param name Base for file name to use for new vignette. Should consist only
 #'   of numbers, letters, _ and -. I recommend using lower case.
 #' @param title The title of the vignette.
@@ -19,29 +22,44 @@
 #' }
 use_vignette <- function(name, title = name) {
   check_is_package("use_vignette()")
-  stopifnot(is_string(name))
-  stopifnot(is_string(title))
-  check_installed("rmarkdown")
 
   use_dependency("knitr", "Suggests")
   use_description_field("VignetteBuilder", "knitr")
-  use_dependency("rmarkdown", "Suggests")
+  use_git_ignore("inst/doc")
+
+  use_vignette_template("vignette.Rmd", name, title)
+}
+
+#' @export
+#' @rdname use_vignette
+use_article <- function(name, title = name) {
+  check_is_package("use_article()")
+
+  path <- use_vignette_template("article.Rmd", name, title)
+  use_build_ignore(path)
+
+  invisible()
+}
+
+use_vignette_template <- function(template, name, title) {
+  stopifnot(is_string(name))
+  stopifnot(is_string(title))
 
   use_directory("vignettes")
   use_git_ignore(c("*.html", "*.R"), directory = "vignettes")
-  use_git_ignore("inst/doc")
+  use_dependency("rmarkdown", "Suggests")
 
-  path <- proj_path("vignettes", asciify(name), ext = "Rmd")
-
-  done("Creating {value(proj_rel_path(path))}")
+  path <- path("vignettes", asciify(name), ext = "Rmd")
 
   data <- project_data()
   data$vignette_title <- title
   data$braced_vignette_title <- glue::glue("{{{title}}}")
 
-  use_template("vignette.Rmd",
-    save_as = proj_rel_path(path),
-    data = data
+  use_template(template,
+    save_as = path,
+    data = data,
+    open = TRUE
   )
-  edit_file(path)
+
+  path
 }